Pulled out AbstractNonBlockingSocketReader
this class can be shared between client and server implementations by subclassing.
